refactored the backend code
